chat: Address Agent Host file read review
Share missing file-read log suppression logic across protocol client and server paths, and avoid assuming read failures are always Error instances.

(Written by Copilot)
